+++ programs/drift/src/state/perp_market.rs
@@ -948,6 +948,7 @@ impl PerpMarket {
             safe_oracle_price_data.delay,
             clock_slot,
             fill_mode.is_liquidation(),
+            user_can_skip_auction_duration,
         )?;
 
         // Proceed if order is low risk and we can fill it. Otherwise check if we can higher risk order immediately
@@ -1448,6 +1449,7 @@ impl AMM {
         ))
     }
 
+    // direction with_amm is the net user direction
     pub fn get_protocol_owned_position(self) -> DriftResult<i64> {
         self.base_asset_amount_with_amm.cast::<i64>()
     }
@@ -1455,15 +1457,18 @@ impl AMM {
     pub fn get_max_reference_price_offset(self) -> DriftResult<i64> {
         if self.curve_update_intensity <= 100 {
             return Ok(0);
+        } else if self.curve_update_intensity >= 200 {
+            // mimic old max behavior with 100 bps
+            return Ok((self.max_spread.cast::<i64>()? / 2).max(10_000));
         }
 
         let lower_bound_multiplier: i64 =
             self.curve_update_intensity.safe_sub(100)?.cast::<i64>()?;
 
-        // always higher of 1-100 bps of price offset and half of the market's max_spread
+        // always the lesser of 1-100 bps of price offset and half of the market's max_spread
         let lb_bps =
             (PERCENTAGE_PRECISION.cast::<i64>()? / 10000).safe_mul(lower_bound_multiplier)?;
-        let max_offset = (self.max_spread.cast::<i64>()? / 2).max(lb_bps);
+        let max_offset = (self.max_spread.cast::<i64>()? / 2).min(lb_bps);
 
         Ok(max_offset)
     }