+++ CHANGELOG.md
@@ -9,6 +9,46 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- program: stricter logic for atomic fills [#2042](https://github.com/drift-labs/protocol-v2/pull/2042)
+
+### Fixes
+
+### Breaking
+
+## [2.149.1] - 2025-11-19
+
+### Features
+
+### Fixes
+
+### Breaking
+- program: add DepositRecord::spot_balance_after [#2034](https://github.com/drift-labs/protocol-v2/pull/2034)
+
+## [2.149.0] - 2025-11-19
+
+### Features
+- sdk: allow deposit from external authority directly to drift account
+
+### Fixes
+
+### Breaking
+
+## [2.148.0] - 2025-11-14
+
+### Features
+
+- pyth lazer zero guard [#2023](https://github.com/drift-labs/protocol-v2/pull/2023)
+
+### Fixes
+
+### Breaking
+
+## [2.147.0] - 2025-11-06
+
+### Features
+
+- whitelist ATA program in begin_swap ix [#2021](https://github.com/drift-labs/protocol-v2/pull/2021)
+
 ### Fixes
 
 ### Breaking